+++ dokumentaatio/changelog.md
@@ -29,3 +29,4 @@
 - Custom Game -pelimuotoa varten on lisätty liukusäätimet, joiden suunnitteluun sekä implementaatioon hyödynnetty tekoälyä (GitHub Copilot)
 - Muutettu Board-luokassa oleva miinanpaljastusalgoritmi rekursiivisesta iteratiiviseksi
 - Lisätty sqlite3 -tietokanta pelaajan highscoren tallentamiseksi/näyttämiseksi
+- Poistettu GameConfig-luokka ja lisätty sen sisältö GameUI-luokkaan